Prefix prompt with : and suffix with ;
By prefixing the prompt with ":" and suffixing with ";", lines from
terminal scrollbacks can be copied and pasted directly into terminals
and the command re-executed.

Thanks @thingskatedid for the idea for ";", taken from the rc shell

":" added so we can continue to have the nice prompt stuff prior to
the ";" (hostname, colours, etc.), which are all treated as command
line arguments to the ":" program, and ignored.

Use ANSI escape sequence to draw the ";" as hidden. Disguise the ";"
by setting the same background and foreground colours.